feat: minor changes in ui.theme
