+++ samples/sample-kotlin-android/src/main/kotlin/com/rudderstack/android/sampleapp/ui/theme/Color.kt
